chore: refactor setServerAuth to supabaseServerClient
