+++ src/nextjs/index.ts
@@ -5,5 +5,5 @@ export { User } from '@supabase/supabase-js';
 export * from './handlers';
 export { default as getUser } from './utils/getUser';
 export { default as withAuthRequired } from './utils/withAuthRequired';
-export { default as setServerAuth } from './utils/setServerAuth';
+export { default as supabaseServerClient } from './utils/supabaseServerClient';
 export { supabaseClient, SupabaseClient } from './utils/initSupabase';
