+++ src/nextjs/utils/supabaseServerClient.ts
@@ -1,22 +1,16 @@
 import { GetServerSidePropsContext, NextApiRequest } from 'next';
-import { SupabaseClient } from '@supabase/supabase-js';
+import { supabaseClient, SupabaseClient } from './initSupabase';
 import { CookieOptions } from '../types';
 
 /**
  * This is a helper method to wrap your SupabaseClient to inject a user's access_token to make use of RLS on the server side.
  *
  * ```js
- * import { setServerAuth } from '@supabase/supabase-auth-helpers/nextjs';
- * import { createClient } from '@supabase/supabase-js';
- *
- * const supabaseClient = createClient(
- *   process.env.NEXT_PUBLIC_SUPABASE_URL!,
- *   process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
- * );
+ * import { supabaseServerClient } from '@supabase/supabase-auth-helpers/nextjs';
  *
  * export async function getServerSideProps(context) {
  *   // Run queries with RLS on the server
- *   const { data } = await setServerAuth(supabaseClient, context)
+ *   const { data } = await supabaseServerClient(context)
  *     .from('test')
  *     .select('*');
  *   return {
@@ -33,8 +27,7 @@ import { CookieOptions } from '../types';
  * @category Server
  */
 
-export default function setServerAuth(
-  supabaseClient: SupabaseClient,
+export default function supabaseServerClient(
   context: GetServerSidePropsContext | { req: NextApiRequest },
   cookieOptions: CookieOptions = {
     name: 'sb'
